break out main toolbar
